1 Translate the UserTest class to English.
